requirements.txt
# GB Power Market JJ - Python Dependencies
# BESS Dashboard & Energy Analysis System

# Core Google APIs
gspread==6.2.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.147.0
google-cloud-bigquery==3.25.0
oauth2client==4.1.3

# Google Sheets Formatting
gspread-formatting==1.2.1

# Data Processing
pandas==2.2.3
numpy==2.1.3

# Visualization
matplotlib==3.9.2
seaborn==0.13.2

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
